modify: 데이터 타입 변경

+++ libs/ml_core/model_contract.py
@@ -38,15 +38,22 @@ RETURN_FEATURE_COLUMNS = [*common_config.BASE_FEATURE_COLUMNS, "return_lag_1h"]
 # (feature_engine/spark/build_merged_table.py에서 이관 — 근거/실측 범위는 그 파일 참고).
 # bike_count/stockout_flag/minute처럼 FEATURE_COLUMNS엔 없지만 feature_engine 내부
 # 중간 산출물에 쓰이는 컬럼도 포함한다(build_merged_table.py가 그대로 재사용).
-# day는 2000-01-01 기준 경과일수(ml_core.day_index) — 부호 비트가 필요 없어
-# int16과 같은 2바이트 비용으로 65,535까지(int16은 32,767까지) 표현하는
-# uint16으로 둔다.
+# day(2000-01-01 기준 경과일수, ml_core.day_index)/station_no는 int16 — Spark엔
+# unsigned 타입이 없어(ShortType이 최대) build_merged_table.py가 이미 부호 있는
+# 16비트로 저장한다. pandas 쪽만 uint16으로 선언해도 Spark가 쓴 값 자체가 이미
+# int16 범위(32,767)로 잘려 있어 실제로 더 넓은 범위를 못 쓴다 — 그래서 여기도
+# 있는 그대로 int16으로 맞춘다(32,767까지, day 기준 약 서기 2089년까지 충분).
 NATIVE_COLUMN_DTYPES = {
     "bike_count": "int16",
     "stockout_flag": "int8",
     "rental_count": "int16",
     "return_count": "int16",
-    "capacity": "float32",
+    # capacity(거치대 수) — 예전엔 "LEFT JOIN이라 결측 가능"이라 float32였는데,
+    # 실제로 이 join은 결측을 낼 수 없다는 걸 코드 경로로 확인했다(2026-08,
+    # build_merged_table.py의 NATIVE_COLUMN_DTYPES 주석에 join 체인 근거 전체 기록).
+    # 항상 정수(거치대 개수)라 int16으로 줄인다 — station_no/day와 이유가 다름을
+    # 표시하기 위해 별도 주석으로 남긴다.
+    "capacity": "int16",
     "lat": "float32",
     "lon": "float32",
     "temp": "float32",
@@ -59,17 +66,29 @@ NATIVE_COLUMN_DTYPES = {
     "dow": "int8",
     "is_holiday": "int8",
     "horizon": "int8",  # 1~HORIZON_COUNT(기본 12) — common_config.HORIZON_COUNT 참고
-    "day": "uint16",
-    # station_no(station_id 대신 모델이 실제로 쓰는 정류소 카테고리 키)도 day와
-    # 같은 이유로 uint16 — 카테고리 코드 자체는 load_station_dtype()이 별도 관리한다
-    # (아래 _feature_column_dtypes()가 station_no는 이 dtype 매핑에서 빼고 카테고리로만
-    # 다룬다 — station_id를 빼던 것과 같은 이유).
-    "station_no": "uint16",
+    "day": "int16",
+    # station_no(station_id 대신 모델이 실제로 쓰는 정류소 카테고리 키) — 카테고리
+    # 코드 자체는 load_station_dtype()이 별도 관리한다(아래 _feature_column_dtypes()가
+    # station_no는 이 dtype 매핑에서 빼고 카테고리로만 다룬다 — station_id를 빼던
+    # 것과 같은 이유).
+    "station_no": "int16",
     # int16이 아니라 float32인 이유: 둘 다 결측 가능하다 — 배치(feature_engine)는
     # LEFT JOIN 결과가 그리드 구멍에서 null일 수 있고, 서빙(predict_single.py)은
     # profile fallback마저 실패하면 np.nan을 그대로 채운다. pandas의 plain int16은
     # NaN을 표현할 수 없어 `.astype()`에서 바로 크래시한다(LightGBM은 결측을
     # 네이티브로 처리하므로 float32 그대로 두는 게 맞다).
+    #
+    # **pandas nullable "Int16"(대문자, 결측 지원)로 다이어트하는 안을 검토했으나
+    # 기각했다(2026-08)**: LightGBM에 실제로 태워보면 float32-NaN과 Int16-pd.NA가
+    # bit-identical한 예측을 내서 그 자체는 문제 없다 — 그런데 predict_single.py의
+    # `_profile_stat()` fallback(`rental_mean`/`return_mean`, station_hourly_profile의
+    # 표본 평균)은 실측값 평균이라 **거의 항상 소수**다(예: 3.4). `Series.astype("Int16")`은
+    # 그런 소수 float에 대해 `TypeError: cannot safely cast non-equivalent float64 to
+    # int16`으로 바로 죽는다(실측 확인 완료) — 배치(feature_engine) 쪽은 fallback 없이
+    # 항상 정수-또는-null이라 문제 없지만, 서빙의 fallback 경로에서 그대로 크래시한다.
+    # fallback을 반올림해서 정수로 강제하면 가능하지만, 그건 "평균값의 소수 정밀도를
+    # 버린다"는 별도의 동작 변경이라 여기서 마음대로 결정하지 않았다 — station_no/day/
+    # capacity와 달리 이 둘은 float32로 남긴다.
     "rental_lag_1h": "float32",
     "return_lag_1h": "float32",
 }
